+++ executor_nocoverage.go
@@ -1,3 +1,4 @@
 package dogma
 
 func (IdempotencyKeyOption) isExecuteCommandOption() {}
+func (EventObserverOption) isExecuteCommandOption()  {}